// frontend/src/api/models/createManualBrokerageAccountDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Splice API
 * Financial account management API
 * OpenAPI spec version: 1.0
 */
import type { ManualBrokeragePositionInput } from './manualBrokeragePositionInput'

export interface CreateManualBrokerageAccountDto {
  /** @minLength 1 */
  name: string
  /** @nullable */
  customName?: string | null
  /** @nullable */
  notes?: string | null
  /** @pattern ^[A-Z]{3}$ */
  accountCurrency: string
  /** @minItems 1 */
  positions: ManualBrokeragePositionInput[]
}