// frontend/src/api/models/replaceManualBrokerageHoldingsDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Splice API
 * Financial account management API
 * OpenAPI spec version: 1.0
 */
import type { ManualBrokeragePositionInput } from './manualBrokeragePositionInput'

export interface ReplaceManualBrokerageHoldingsDto {
  positions: ManualBrokeragePositionInput[]
}